Fix issue where AsyncStorage can't re-set unsentEventString nor unsentIdentifyKey when they are null
As part of migrating unsentEvents, we are appending this._storageSuffix to both the unsentKey and unsentIdentifyKey.
When either one of these is null, AsyncStorage.setItem breaks because it expects strings.
To fix this, I'm only migrating these values if they are valid and not null. It might be an overkill, but I'm
also explicitly JSON.stringifying the values.